backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
PyPDF2==3.0.1
pdfplumber==0.10.0
langchain==0.1.0
openai==1.3.0
chromadb==0.4.18
pydantic-settings==2.1.0
python-dotenv==1.0.0 